Move IDependenciesRuleHandler
This interface is not specific to "cross targeting". It's more closely aligned with the subscriptions, and is implemented within a subnamespace of the target namespace.
